Create constants folder
