+++ packages/twenty-server/src/modules/workflow/workflow-trigger/automated-trigger/constants/automated-trigger-settings.ts
@@ -1,0 +1,19 @@
+export type BaseDatabaseEventTriggerSettings = {
+  eventName: string;
+};
+
+export type DatabaseEventTriggerSettings =
+  | BaseDatabaseEventTriggerSettings
+  | UpdateEventTriggerSettings;
+
+export type UpdateEventTriggerSettings = BaseDatabaseEventTriggerSettings & {
+  fields: string[];
+};
+
+export type CronTriggerSettings = {
+  pattern: string;
+};
+
+export type AutomatedTriggerSettings =
+  | DatabaseEventTriggerSettings
+  | CronTriggerSettings;
